Fixed a bug in xdot.py because it seems that coordinates can be float and not int.

+++ lib/ui/xdot.py
@@ -1065,7 +1065,7 @@ class XDotParser(DotParser):
             if not bb:
                 return
 
-            xmin, ymin, xmax, ymax = map(int, bb.split(","))
+            xmin, ymin, xmax, ymax = map(float, bb.split(","))
 
             self.xoffset = -xmin
             self.yoffset = -ymax
